refactor(aigateway): split the OpenRouter dispatch hardening out of the plugin
`plugins/openrouter_provider/plugin.py` was 510 lines against the 450-line
limit the plan binds to touched Python files. The overage was a block of
module-level helpers ahead of the plugin class, and it separates along two
distinct concerns rather than an arbitrary line cut:

- `litellm_controls.py` — everything that exists only because OpenRouter
  dispatches through a SHARED LiteLLM surface: the orchestration selectors a
  caller can plant in the request body, and the process-global callbacks and
  routing that can observe or redirect an account-scoped BYOK credential.
  Neither is OpenRouter vocabulary, and both are pure functions over data.
- `dispatch_errors.py` — the sanitized gateway errors raised on the dispatch
  path. Deliberately separate from `response_errors.py`: that module DETECTS
  an error inside a provider payload, this one CONSTRUCTS the gateway's answer
  to one. Keeping them apart is what stops a detector from gaining the ability
  to shape the client-facing error.

Behaviour-preserving, and proved rather than asserted:

- Zero test files touched — the append-only gate passing unskipped is the
  machine proof.
- pyright green: every import site still resolves.
- Module surface diffed by AST against the previous revision. Nothing became
  newly reachable through `plugin`; the names that left are three incidental
  import leaks (`HTTPException`, `Mapping`, `NonRetryableProviderError`) and
  six private symbols with no reference anywhere outside the package.
  `OFFICIAL_API_BASE`, `PLUGIN`, both public classes and the private
  `_top_level_error_is_meaningful` re-export are all still importable from
  `plugin`.
- `_UnsafeLiteLLMStateError` was the one moved symbol with runtime coupling;
  `core/retry.py` reads its marker duck-typed, not by class identity, so the
  relocation cannot change a retry decision.

382 / 94 / 75 lines.

Refs: OME-653
